# src/main/generated/src/main/java/org/openapi/model/SelectHDStopTransResponsError.java
/*
 * Aristo Stop Transactions
 * Provide Aristo Stop Transactions
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: ravviz@nytimes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package org.openapi.model;

import java.util.Objects;
import java.util.Arrays;
import com.google.gson.TypeAdapter;
import com.google.gson.annotations.JsonAdapter;
import com.google.gson.annotations.SerializedName;
import com.google.gson.stream.JsonReader;
import com.google.gson.stream.JsonWriter;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.io.IOException;

/**
 * SelectHDStopTransResponsError
 */
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2022-02-01T15:08:08.748301-05:00[America/New_York]")
public class SelectHDStopTransResponsError {
  public static final String SERIALIZED_NAME_PRODUCT_ID = "product_id";
  @SerializedName(SERIALIZED_NAME_PRODUCT_ID)
  private String productId;

  public static final String SERIALIZED_NAME_ERROR_CODE = "error_code";
  @SerializedName(SERIALIZED_NAME_ERROR_CODE)
  private Integer errorCode;

  public static final String SERIALIZED_NAME_ERROR_MSG = "error_msg";
  @SerializedName(SERIALIZED_NAME_ERROR_MSG)
  private String errorMsg;

  public static final String SERIALIZED_NAME_ORIGINAL_TRANSACTION_ID = "original_transaction_id";
  @SerializedName(SERIALIZED_NAME_ORIGINAL_TRANSACTION_ID)
  private String originalTransactionId;


  public SelectHDStopTransResponsError productId(String productId) {
    
    this.productId = productId;
    return this;
  }

   /**
   * Get productId
   * @return productId
  **/
  @javax.annotation.Nonnull
  @ApiModelProperty(required = true, value = "")

  public String getProductId() {
    return productId;
  }


  public void setProductId(String productId) {
    this.productId = productId;
  }


  public SelectHDStopTransResponsError errorCode(Integer errorCode) {
    
    this.errorCode = errorCode;
    return this;
  }

   /**
   * Get errorCode
   * @return errorCode
  **/
  @javax.annotation.Nonnull
  @ApiModelProperty(required = true, value = "")

  public Integer getErrorCode() {
    return errorCode;
  }


  public void setErrorCode(Integer errorCode) {
    this.errorCode = errorCode;
  }


  public SelectHDStopTransResponsError errorMsg(String errorMsg) {
    
    this.errorMsg = errorMsg;
    return this;
  }

   /**
   * Get errorMsg
   * @return errorMsg
  **/
  @javax.annotation.Nonnull
  @ApiModelProperty(required = true, value = "")

  public String getErrorMsg() {
    return errorMsg;
  }


  public void setErrorMsg(String errorMsg) {
    this.errorMsg = errorMsg;
  }


  public SelectHDStopTransResponsError originalTransactionId(String originalTransactionId) {
    
    this.originalTransactionId = originalTransactionId;
    return this;
  }

   /**
   * Get originalTransactionId
   * @return originalTransactionId
  **/
  @javax.annotation.Nonnull
  @ApiModelProperty(required = true, value = "")

  public String getOriginalTransactionId() {
    return originalTransactionId;
  }


  public void setOriginalTransactionId(String originalTransactionId) {
    this.originalTransactionId = originalTransactionId;
  }


  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    SelectHDStopTransResponsError selectHDStopTransResponsError = (SelectHDStopTransResponsError) o;
    return Objects.equals(this.productId, selectHDStopTransResponsError.productId) &&
        Objects.equals(this.errorCode, selectHDStopTransResponsError.errorCode) &&
        Objects.equals(this.errorMsg, selectHDStopTransResponsError.errorMsg) &&
        Objects.equals(this.originalTransactionId, selectHDStopTransResponsError.originalTransactionId);
  }

  @Override
  public int hashCode() {
    return Objects.hash(productId, errorCode, errorMsg, originalTransactionId);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class SelectHDStopTransResponsError {\n");
    sb.append("    productId: ").append(toIndentedString(productId)).append("\n");
    sb.append("    errorCode: ").append(toIndentedString(errorCode)).append("\n");
    sb.append("    errorMsg: ").append(toIndentedString(errorMsg)).append("\n");
    sb.append("    originalTransactionId: ").append(toIndentedString(originalTransactionId)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }

}