# src/main/generated/src/main/java/org/openapi/model/SelectHDStopTransResponse.java
/*
 * Aristo Stop Transactions
 * Provide Aristo Stop Transactions
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: ravviz@nytimes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package org.openapi.model;

import java.util.Objects;
import java.util.Arrays;
import com.google.gson.TypeAdapter;
import com.google.gson.annotations.JsonAdapter;
import com.google.gson.annotations.SerializedName;
import com.google.gson.stream.JsonReader;
import com.google.gson.stream.JsonWriter;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.io.IOException;
import java.util.ArrayList;
import java.util.List;
import org.openapi.model.SelectHDStopTransResponseData;

/**
 * SelectHDStopTransResponse
 */
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2022-02-01T15:08:08.748301-05:00[America/New_York]")
public class SelectHDStopTransResponse {
  public static final String SERIALIZED_NAME_TRANSACTIONS = "transactions";
  @SerializedName(SERIALIZED_NAME_TRANSACTIONS)
  private List<SelectHDStopTransResponseData> transactions = new ArrayList<SelectHDStopTransResponseData>();


  public SelectHDStopTransResponse transactions(List<SelectHDStopTransResponseData> transactions) {
    
    this.transactions = transactions;
    return this;
  }

  public SelectHDStopTransResponse addTransactionsItem(SelectHDStopTransResponseData transactionsItem) {
    this.transactions.add(transactionsItem);
    return this;
  }

   /**
   * Get transactions
   * @return transactions
  **/
  @javax.annotation.Nonnull
  @ApiModelProperty(required = true, value = "")

  public List<SelectHDStopTransResponseData> getTransactions() {
    return transactions;
  }


  public void setTransactions(List<SelectHDStopTransResponseData> transactions) {
    this.transactions = transactions;
  }


  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    SelectHDStopTransResponse selectHDStopTransResponse = (SelectHDStopTransResponse) o;
    return Objects.equals(this.transactions, selectHDStopTransResponse.transactions);
  }

  @Override
  public int hashCode() {
    return Objects.hash(transactions);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class SelectHDStopTransResponse {\n");
    sb.append("    transactions: ").append(toIndentedString(transactions)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }

}