# src/main/generated/src/main/java/org/openapi/model/SelectHDStopTransResponseData.java
/*
 * Aristo Stop Transactions
 * Provide Aristo Stop Transactions
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: ravviz@nytimes.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package org.openapi.model;

import java.util.Objects;
import java.util.Arrays;
import com.google.gson.TypeAdapter;
import com.google.gson.annotations.JsonAdapter;
import com.google.gson.annotations.SerializedName;
import com.google.gson.stream.JsonReader;
import com.google.gson.stream.JsonWriter;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import java.io.IOException;
import java.util.ArrayList;
import java.util.List;
import org.openapi.model.SelectHDStopTransResponsError;

/**
 * SelectHDStopTransResponseData
 */
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2022-02-01T15:08:08.748301-05:00[America/New_York]")
public class SelectHDStopTransResponseData {
  public static final String SERIALIZED_NAME_ERRORS = "errors";
  @SerializedName(SERIALIZED_NAME_ERRORS)
  private List<SelectHDStopTransResponsError> errors = null;


  public SelectHDStopTransResponseData errors(List<SelectHDStopTransResponsError> errors) {
    
    this.errors = errors;
    return this;
  }

  public SelectHDStopTransResponseData addErrorsItem(SelectHDStopTransResponsError errorsItem) {
    if (this.errors == null) {
      this.errors = new ArrayList<SelectHDStopTransResponsError>();
    }
    this.errors.add(errorsItem);
    return this;
  }

   /**
   * Get errors
   * @return errors
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "")

  public List<SelectHDStopTransResponsError> getErrors() {
    return errors;
  }


  public void setErrors(List<SelectHDStopTransResponsError> errors) {
    this.errors = errors;
  }


  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    SelectHDStopTransResponseData selectHDStopTransResponseData = (SelectHDStopTransResponseData) o;
    return Objects.equals(this.errors, selectHDStopTransResponseData.errors);
  }

  @Override
  public int hashCode() {
    return Objects.hash(errors);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class SelectHDStopTransResponseData {\n");
    sb.append("    errors: ").append(toIndentedString(errors)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }

}